requirements.txt
torchvision>=0.19.0
scipy>=1.14.1
tensorboard>=2.17